+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# ps development version
+
 # ps 1.4.0
 
 * ps is now under the MIT license.