+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# ps development version
+# ps 1.5.0
 
 * New function `ps_shared_libs()` to list the loaded shared libraries
   of a process, on Windows.